Refactor.
A.k.a, "Enough, already. Let slip the dogs of ReSharper..."
This also gets HTTP/2 working.  It can try HTTP/3, but the connection is always
rejected (see the Http3Support flag in the .csproj file).
Split the mega-folder into smaller folders. Put things into subdirectories
based on the last three characters of the filename.  This spreads the
directory updates over multiple directories (since we fetch in order).